+++ CHANGES.md
@@ -7,12 +7,13 @@
   - Birthday
   - Holiday
   - Task
-- MAJOR: Added side menu for the whole Calendar, which allows Groups and Event Types to be toggled easily.
+- MAJOR: Added side menu for the whole Calendar, which allows Visible Days, Event Types, and Groups to be toggled easily.
 - MAJOR: Updating configuration will no longer close your current view (All Events, Full Day, Weekly).  They will be automatically updated.
 - Added new public functions "addEventType()" and "removeEventType()", for adding adding and removing event types.
 - Added new public function "setVisibleGroups()", which will set the group name that are visible.
 - Added new public function "setVisibleEventTypes()", which will set the event types that are visible.
 - Removed the "Groups" tab from the "Configuration" dialog (as its all in the side menu now).
+- Removed the "Visible Days" tab from the "Configuration" dialog (as its all in the side menu now).
 - The "View Current Week Events" button is now the same, but flipped.
 - When events are added/updated, the group they are assigned (if new) is made visible by default.
 - The "locked" property is now included in the CSV export result.